// projects/image-forge/src/lib/models/Layers/LayerBlendModes.ts
export enum LayerBlendModes {
  SourceOver = 'source-over',
  SourceIn = 'source-in',
  SourceOut = 'source-out',
  SourceAtop = 'source-atop',
  DestinationOver = 'destionation-over',
  DestinationIn = 'destination-in',
  DestinationOut = 'destination-out',
  DestinationAtop = 'destination-atop',
  Lighter = 'lighter',
  Copy = 'copy',
  Xor = 'xor',
  Multiply = 'multiply',
  Screen = 'screen',
  Overlay = 'overlay',
  Darken = 'darken',
  Lighten = 'lighten',
  ColorDodge = 'color-dodge',
  ColorBurn = 'color-burn',
  HardLight = 'hard-light',
  SoftLight = 'soft-light',
  Difference = 'difference',
  Exclusion = 'exclusion',
  Hue = 'hue',
  Saturation = 'saturation',
  Color = 'color',
  Luminosity = 'luminosity',
}